qcluster detection start
started working on a way to detect if qcluster is active or not. If not it should throw an error. currently cluster_is_running function is set to True. Should update it and actually detect if cluster is running. Stat or cluster is not working to detect that, becuase it isnt in current django-q version. should upgrade it if possible. committing before upgrading in case it messes anything up.